added support for readonly mode

+++ eneel/adapters/postgres.py
@@ -9,7 +9,7 @@ logger = logging.getLogger('main_logger')
 
 
 class Database:
-    def __init__(self, server, user, password, database, limit_rows=None):
+    def __init__(self, server, user, password, database, limit_rows=None, read_only=False):
         try:
             conn_string = "host=" + server + " dbname=" + \
                           database + " user=" + user + " password=" + password
@@ -19,6 +19,7 @@ class Database:
             self._database = database
             self._dialect = "postgres"
             self._limit_rows = limit_rows
+            self._read_only = read_only
 
             self._conn = psycopg2.connect(conn_string)
             self._conn.autocommit = True
@@ -150,6 +151,8 @@ class Database:
             logger.error("Failed checking table exist")
 
     def truncate_table(self, table_name):
+        if self._read_only:
+            sys.exit("This source is readonly. Terminating load run")
         try:
             sql = "TRUNCATE TABLE " + table_name
             self.execute(sql)
@@ -158,6 +161,8 @@ class Database:
             logger.error("Failed truncating table")
 
     def create_schema(self, schema):
+        if self._read_only:
+            sys.exit("This source is readonly. Terminating load run")
         try:
             if schema in self.schemas():
                 logger.debug("Schema exists")
@@ -225,6 +230,8 @@ class Database:
             logger.error("Failed exporting table")
 
     def insert_from_table_and_drop(self, schema, to_table, from_table):
+        if self._read_only:
+            sys.exit("This source is readonly. Terminating load run")
         to_schema_table = schema + "." + to_table
         from_schema_table = schema + "." + from_table
         try:
@@ -234,6 +241,8 @@ class Database:
             logger.error("Failed to insert_from_table_and_drop")
 
     def switch_tables(self, schema, old_table, new_table):
+        if self._read_only:
+            sys.exit("This source is readonly. Terminating load run")
         try:
 
             old_schema_table = schema + "." + old_table
@@ -253,6 +262,8 @@ class Database:
             logger.error("Failed to switch tables")
 
     def import_table(self, schema, table, file, delimiter=','):
+        if self._read_only:
+            sys.exit("This source is readonly. Terminating load run")
         try:
             schema_table = schema + '.' + table
 
@@ -310,6 +321,8 @@ class Database:
             logger.error("Failed generating create table script")
 
     def create_table_from_columns(self, schema, table, columns):
+        if self._read_only:
+            sys.exit("This source is readonly. Terminating load run")
         try:
             table_exists = self.query("SELECT * FROM INFORMATION_SCHEMA.TABLES "
                                         "WHERE TABLE_SCHEMA = %s AND "
